moved tests to their own file
